# api/openapi.yaml
openapi: 3.0.3
info:
  title: Yggdrasil Control Plane API
  version: 1.4.0
  description: >-
    Local AI control plane API for Yggdrasil, its OpenAI-compatible routes, and its MCP server.
    On the default loopback bind no key is needed. When the daemon listens on any other address,
    every /api/v1, /v1, and /mcp route needs Authorization: Bearer YOUR_API_KEY; /about and /source do not.
    Errors are {"error": {"code", "message", "details"}}. docs/api.md describes the behavior of each route.
servers:
  - url: http://127.0.0.1:7331
security:
  - {}
  - bearerAuth: []
paths:
...
  /api/v1/models/{id}:
    delete:
      summary: Delete a model file
      operationId: deleteModel
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: node_id
          in: query
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                node_id:
                  type: string
      responses:
        "204":
          description: Deleted
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/v1/conversations/{id}/artifacts:
    get:
      summary: Files attached to or made in a conversation
      operationId: listConversationArtifacts
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Files
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Artifact"
...
  /api/v1/images/setup:
    get:
      summary: What image generation has installed, the models on offer, and any setup in progress
      operationId: getImageSetup
      responses:
        "200":
          description: Image generation status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImageSetup"
    post:
      summary: Install stable-diffusion.cpp and an image model in the background
      operationId: startImageSetup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [model_id]
              properties:
                model_id: {type: string, description: "A model from the status's models, such as flux2-klein-4b"}
      responses:
        "202":
          description: Setup started; poll the status for its progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImageSetup"
        "409":
          description: Setup is already running, there is too little space, or this computer cannot run it
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      summary: Stop the setup in progress; what was downloaded is kept and resumed next time
      operationId: cancelImageSetup
      responses:
        "200":
          description: Image generation status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImageSetup"
  /api/v1/images/models/{id}:
    delete:
      summary: Delete an installed image model to free its space
      operationId: removeImageModel
      parameters:
        - name: id
          in: path
          required: true
          schema: {type: string}
      responses:
        "200":
          description: Image generation status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImageSetup"
        "409":
          description: The model is being set up
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/video/setup:
    get:
      summary: What video generation has installed, the models on offer, and any setup in progress
      operationId: getVideoSetup
      responses:
        "200":
          description: Image generation status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImageSetup"
    post:
      summary: Install stable-diffusion.cpp and a video model in the background
      operationId: startVideoSetup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [model_id]
              properties:
                model_id: {type: string, description: "A model from the status's models, such as wan2.2-ti2v-5b"}
      responses:
        "202":
          description: Setup started; poll the status for its progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImageSetup"
        "409":
          description: Setup is already running, there is too little space, or this computer cannot run it
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      summary: Stop the setup in progress; what was downloaded is kept and resumed next time
      operationId: cancelVideoSetup
      responses:
        "200":
          description: Image generation status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImageSetup"
  /api/v1/video/models/{id}:
    delete:
      summary: Delete an installed video model to free its space
      operationId: removeVideoModel
      parameters:
        - name: id
          in: path
          required: true
          schema: {type: string}
      responses:
        "200":
          description: Image generation status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImageSetup"
        "409":
          description: The model is being set up
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/artifacts:
    post:
      summary: Upload a file to attach to a chat (up to 25 MB)
      operationId: uploadArtifact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                text:
                  type: string
                content_base64:
                  type: string
                conversation_id:
                  type: string
              required: [name]
      responses:
        "200":
          description: The file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artifact"
        "400":
          description: Unsupported, unreadable, or invalid file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/artifacts/{id}:
    get:
      summary: One file
      operationId: getArtifact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artifact"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      summary: Delete a file
      operationId: deleteArtifact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/artifacts/{id}/content:
    get:
      summary: A file's content
      operationId: getArtifactContent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: inline
          in: query
          description: "1 shows a PDF or image in the browser instead of downloading it"
          schema:
            type: string
            enum: ["1"]
      responses:
        "200":
          description: The file, with its own content type
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: An API key from the API Access page, or YGGDRASIL_API_KEY
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
...
    ImageSetup:
      type: object
      properties:
        supported: {type: boolean, description: "False where stable-diffusion.cpp has no build or cannot run"}
        unsupported: {type: string}
        ready: {type: boolean, description: "Images can be made now"}
        program: {type: boolean, description: "stable-diffusion.cpp is installed"}
        release: {type: string}
        active: {type: string, description: "The model images are made with"}
        models:
          type: array
          items:
            type: object
            properties:
              id: {type: string}
              name: {type: string}
              description: {type: string}
              license: {type: string}
              memory_bytes: {type: integer, format: int64}
              size_bytes: {type: integer, format: int64}
              edits: {type: boolean, description: "Edits images, or for a video model, animates them"}
              kind: {type: string, enum: [image, video]}
              installed: {type: boolean}
              recommended: {type: boolean}
              files:
                type: array
                items:
                  type: object
                  properties:
                    role: {type: string, enum: [diffusion, vae, llm, t5xxl]}
                    repo: {type: string}
                    path: {type: string}
                    size_bytes: {type: integer, format: int64}
                    sha256: {type: string}
        job:
          type: object
          properties:
            model_id: {type: string}
            stage: {type: string, enum: [program, model]}
            done_bytes: {type: integer, format: int64}
            total_bytes: {type: integer, format: int64}
            running: {type: boolean}
            error: {type: string}
    Artifact:
      type: object
      properties:
        id:
          type: string
        conversation_id:
          type: string
        name:
          type: string
        mime_type:
          type: string
        kind:
          type: string
        size_bytes:
          type: integer
        producer:
          type: string
        source_task:
          type: string
      additionalProperties: true
...